backend/requirements.txt
Flask==2.2.3
Flask-CORS==3.0.10
requests==2.28.2
python-dotenv